#342 filter by violation type
