+++ client/lib/common/asset/less/violation/violation-list.less
@@ -2,35 +2,4 @@
 
 .violationList {
 
-    .violationList-copy-url {
-        margin-bottom: @padding-small 0;
-    }
-
-    .violationList-datepicker {
-        display: flex;
-        flex-direction: row;
-    }
-
-    .violationList-filter {
-        margin-bottom: @padding-small;
-    }
-
-    .violationList-btn-group {
-        > :first-child {
-            margin-right: 3px;
-        }
-    }
-
-    .violationList-info {
-        padding: @padding-tiniest @padding-tiny;
-        font-size: @small-font-size;
-        text-align: center;
-        position: fixed;
-        bottom: 0;
-        background: @gray-light;
-        color: @white;
-        z-index: 10;
-        border: 1px solid @white;
-        border-bottom: 0;
-    }
 }
