+++ client/lib/common/src/data/fullstop/fullstop-getter.js
@@ -7,24 +7,13 @@ function getViolation(state, violationId) {
     return violation ? violation.toJS() : false;
 }
 
-function getViolations(state, accounts, resolved) {
+function getViolations(state) {
     let violations = state.get('violations').valueSeq();
 
     if (violations.count() === 0) {
         return [];
     }
 
-    // collect violations from accounts
-    if (accounts) {
-        violations = violations.filter(v => accounts.indexOf(v.get('account_id')) >= 0);
-    }
-
-    // filter by resolution
-    if (resolved === true) {
-        violations = violations.filter(v => v.get('comment') !== null);
-    } else if (resolved === false) {
-        violations = violations.filter(v => v.get('comment') === null);
-    }
     return violations
             .sortBy(v => v.get('timestamp'))
             .toJS();
