+++ package.json
@@ -36,6 +36,7 @@
     "react-gravatar": "2.0.1",
     "react-infinite-scroll": "0.1.5",
     "react-router": "0.13.3",
+    "react-select": "0.6.12",
     "react-tabs": "0.4.1",
     "react-time": "2.5.0",
     "react-typeahead": "1.1.1",
